chech fragments-dir

+++ picture-transform-asp.net_core-react/Program.cs
@@ -1,4 +1,7 @@
-Directory.Delete(Path.Combine("ClientApp", "public", "image-fragments"),true); 
+if (Directory.Exists(Path.Combine("ClientApp", "public", "image-fragments")))
+{
+    Directory.Delete(Path.Combine("ClientApp", "public", "image-fragments"),true);
+} 
 
 var builder = WebApplication.CreateBuilder(args);
